backend config  to port 8000, completed LoginFormPage
