+++ frontend/index.html
@@ -2,9 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/x-icon" href="/favicon.ico" />
+    <link rel="icon" type="image/x-icon" href="/bnb-1.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Authenticate Me</title>
+    <title>treeTopBnB</title>
   </head>
   <body>
     <div id="root"></div>
