+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "authenticate-me",
+  "name": "treetopbnb",
   "private": true,
   "version": "0.1.0",
   "type": "module",
