+++ frontend/src/components/LoginFormPage/index.js
@@ -1,0 +1,2 @@
+import LoginFormPage from "./LoginFormPage ";
+export default LoginFormPage;
